requirements.txt
tensorflow-cpu
streamlit